[CI] Consolidate multimodal entrypoint tests. (#44819)
