+++ tests/entrypoints/multimodal/llm/test_mm_cache_external_injection.py
@@ -15,7 +15,7 @@ import logging
 import pytest
 import regex as re
 
-from tests.entrypoints.openai.chat_completion.test_vision import TEST_IMAGE_ASSETS
+from tests.entrypoints.multimodal.conftest import TEST_IMAGE_ASSETS
 from vllm import LLM, SamplingParams
 from vllm.renderers.params import ChatParams
 from vllm.v1.metrics import loggers as stat_loggers
