+++ tests/entrypoints/multimodal/llm/test_mm_embeds_only.py
