+++ tests/entrypoints/multimodal/llm/test_mm_processor_kwargs.py
