+++ tests/entrypoints/multimodal/openai/chat_completion/test_audio.py
