+++ tests/entrypoints/multimodal/openai/chat_completion/test_audio_in_video.py
