+++ tests/entrypoints/multimodal/openai/chat_completion/test_chat_completion_with_image_embeds.py
@@ -52,7 +52,7 @@ async def client_with_image_embeds(server_with_image_embeds):
 @pytest.mark.asyncio
 @pytest.mark.parametrize("model_name", [MODEL_NAME])
 @pytest.mark.parametrize("dtype", [torch.half, torch.float16, torch.float32])
-async def test_completions_with_image_embeds(
+async def test_chat_completions_with_image_embeds(
     client_with_image_embeds: openai.AsyncOpenAI,
     model_name: str,
     image_assets: ImageTestAssets,
