+++ tests/entrypoints/multimodal/openai/chat_completion/test_default_mm_loras.py
