+++ tests/entrypoints/multimodal/openai/chat_completion/test_video.py
