+++ tests/entrypoints/multimodal/openai/chat_completion/test_vision_embeds.py
